ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
pillow==10.0.1
python-multipart==0.0.6
